Module Cohttp

module Accept : sig ... end

Accept-Encoding HTTP header parsing and generation

module Auth : sig ... end

HTTP Authentication and Authorization header parsing and generation

module Body : sig ... end

HTTP request and response body handling

module Conf : sig ... end

Compile-time configuration variables

module Connection : sig ... end

module Code : sig ... end

module Cookie : sig ... end

Functions for the HTTP Cookie and Set-Cookie header fields. Using the Set-Cookie header field, an HTTP server can pass name/value pairs and associated metadata (called cookies) to a user agent. When the user agent makes subsequent requests to the server, the user agent uses the metadata and other information to determine whether to return the name/value pairs in the Cookie header.

module Header : sig ... end

module Link : sig ... end

RFC 5988 ("Web Linking") and RFC 5987 ("Character Set and Language Encoding for Hypertext Transfer Protocol (HTTP) Header Field Parameters")

module Request : sig ... end

HTTP/1.1 request handling

module Response : sig ... end

HTTP/1.1 response handling

module S : sig ... end

Module type signatures for Cohttp components

module Path : sig ... end

module Transfer : sig ... end

Read and write the HTTP/1.1 transfer-encoding formats. Currently supported are chunked and content-length.

module Private : sig ... end