Module Cookie.Set_cookie_hdr

type t = {

cookie : cookie;

expiration : expiration;

domain : string option;

path : string option;

secure : bool;

http_only : bool;

}

A header which a server sends to a client to request that the client returns the cookie in future requests, under certain conditions.

include Ppx_sexp_conv_lib.Sexpable.S with type t := t

val t_of_sexp : Sexplib0.Sexp.t -> t

val sexp_of_t : t -> Sexplib0.Sexp.t

val make : ?⁠expiration:expiration -> ?⁠path:string -> ?⁠domain:string -> ?⁠secure:bool -> ?⁠http_only:bool -> cookie -> t

val serialize : ?⁠version:[ `HTTP_1_0 | `HTTP_1_1 ] -> t -> string * string

Return an HTTP header

val extract : Header.t -> (string * t) list

Return the list of cookies sent by the server

val cookie : t -> cookie

The name-value binding

val value : t -> string

The value

val expiration : t -> expiration

The expiration

val domain : t -> string option

The domain for which the cookie is valid, if any

val path : t -> string option

The path for which the cookie is valid, if any

val secure : t -> bool

Has the cookie's secure attribute been set?

val http_only : t -> bool