Module Header.Private

val caseless_equal : string -> string -> bool

caseless_equal a b must be equivalent to String.equal (String.lowercase_ascii a) (String.lowercase_ascii b).

val first : t -> (string * string) option

first t is Some (hdr_name, hdr_value), which represents the first header in headers list t. It is None if t is empty.

val move_to_front : t -> string -> t

move_to_front t hdr_name is t with header name hdr_name moved to the front of the headers list t. If the header doesn't exist in t or the header is already at the front, then t is unchanged.