Module Cohttp.Request

HTTP/1.1 request handling

This contains the metadata for a HTTP/1.1 request header, including the headers, version, meth and uri. The body is handled by the separate S module type, as it is dependent on the IO implementation.

include S.Request

type t = {

headers : Header.t;
HTTP request headers

meth : Code.meth;
HTTP request method

scheme : string option;
URI scheme (http or https)

resource : string;
Request path and query

version : Code.version;
HTTP version, usually 1.1

encoding : Transfer.encoding;
transfer encoding of this HTTP request

}

val compare : t -> t -> int

val encoding : t -> Transfer.encoding

val version : t -> Code.version

val resource : t -> string

val scheme : t -> string option

val meth : t -> Code.meth

val headers : t -> Header.t

module Fields : sig ... end

include Ppx_sexp_conv_lib.Sexpable.S with type Request.t := t

val t_of_sexp : Sexplib0__.Sexp.t -> t

val sexp_of_t : t -> Sexplib0__.Sexp.t

val make : ?⁠meth:Code.meth -> ?⁠version:Code.version -> ?⁠encoding:Transfer.encoding -> ?⁠headers:Header.t -> Uri.t -> t

Return true whether the connection should be reused

val is_keep_alive : t -> bool

Return true whether the connection should be reused

val uri : t -> Uri.t

val make_for_client : ?⁠headers:Header.t -> ?⁠chunked:bool -> ?⁠body_length:int64 -> Code.meth -> Uri.t -> t

val pp_hum : Stdlib.Format.formatter -> t -> unit

Human-readable output, used by the toplevel printer

module Make (IO : S.IO) : S.Http_io with type t = t and module IO = IO

Functor to construct the IO-specific HTTP request handling functions