Module type S.Response

type t = {

encoding : Transfer.encoding;
(*

deprecated this field will be removed in the future

*)

headers : Header.t;
(*
response HTTP headers
*)

version : Code.version;
(*
(** HTTP version, usually 1.1 *)
*)

status : Code.status_code;
(*
HTTP status code of the response
*)

flush : bool;
(*

deprecated this field will be removed in the future

*)

}

include Sexplib0.Sexpable.S with type t := t

val t_of_sexp : Sexplib0__.Sexp.t -> t

val sexp_of_t : t -> Sexplib0__.Sexp.t

val encoding : t -> Transfer.encoding

val headers : t -> Header.t

val version : t -> Code.version

val status : t -> Code.status_code

val flush : t -> bool

val compare : t -> t -> int

val make : ?version:Code.version -> ?status:Code.status_code -> ?flush:bool -> ?encoding:Transfer.encoding -> ?headers:Header.t -> unit -> t

make () is a value of t. The default values for the request, if not specified, are: status is `Ok, version is `HTTP_1_1, flush is false and headers is Header.empty. The request encoding value is determined via the Header.get_transfer_encoding function and, if not found, uses the default value Transfer.Chunked.