Module type S.Response

type t = {

encoding : Transfer.encoding;
(*
Transfer encoding of this HTTP response
*)

headers : Header.t;
(*
response HTTP headers
*)

version : Code.version;
(*
(** HTTP version, usually 1.1 *)
*)

status : Code.status_code;
(*
HTTP status code of the response
*)

flush : bool;

}

include Sexplib0.Sexpable.S with type t := t

val t_of_sexp : Sexplib0.Sexp.t -> t

val sexp_of_t : t -> Sexplib0.Sexp.t

val encoding : t -> Transfer.encoding

val headers : t -> Header.t

val version : t -> Code.version

val status : t -> Code.status_code

val flush : t -> bool

val compare : t -> t -> int

val make : ?version:Code.version -> ?status:Code.status_code -> ?flush:bool -> ?encoding:Transfer.encoding -> ?headers:Header.t -> unit -> t

The response creates by make ~encoding ~headers () has an encoding value determined from the content of headers or if no proper header is present, using the value of encoding. Checked headers are "content-length", "content-range" and "transfer-encoding". The default value of encoding is chunked.